audio: add variable for SR recorder handle
The ESP-SR models are mmap'd during init. Destroying the audio recorder
that uses the SR recorder does not destroy the SR recorder, and thus the
models are still mmap'd afterwards. This could cause problems if we want
to update the model partition at runtime.

Add a variable for the SR recorder handle so we can destroy it during
audio deinit. We will need this if we want to OTA update the models.